Refactor runtime initialization into separate modules
