+++ crates/runtime/src/init/extension.rs
@@ -1,0 +1,28 @@
+/*
+Copyright 2024 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+use crate::Runtime;
+
+impl Runtime {
+    pub(crate) async fn start_extensions(&self) {
+        let mut extensions = self.extensions.write().await;
+        for (name, extension) in extensions.iter_mut() {
+            if let Err(err) = extension.on_start(self).await {
+                tracing::warn!("Failed to start extension {name}: {err}");
+            }
+        }
+    }
+}
