+++ crates/runtime/src/init/results_cache.rs
@@ -1,0 +1,45 @@
+/*
+Copyright 2024 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+use cache::QueryResultsCacheProvider;
+
+use crate::{datafusion::SPICE_RUNTIME_SCHEMA, Runtime};
+
+impl Runtime {
+    pub async fn init_results_cache(&self) {
+        let app = self.app.read().await;
+        let Some(app) = app.as_ref() else { return };
+
+        let cache_config = &app.runtime.results_cache;
+
+        if !cache_config.enabled {
+            return;
+        }
+
+        match QueryResultsCacheProvider::try_new(
+            cache_config,
+            Box::new([SPICE_RUNTIME_SCHEMA.into(), "information_schema".into()]),
+        ) {
+            Ok(cache_provider) => {
+                tracing::info!("Initialized results cache; {cache_provider}");
+                self.datafusion().set_cache_provider(cache_provider);
+            }
+            Err(e) => {
+                tracing::warn!("Failed to initialize results cache: {e}");
+            }
+        };
+    }
+}
