+++ crates/runtime/src/lib.rs
@@ -15,7 +15,6 @@ limitations under the License.
 */
 #![allow(clippy::missing_errors_doc)]
 
-use std::collections::HashSet;
 use std::net::SocketAddr;
 use std::pin::Pin;
 use std::{collections::HashMap, sync::Arc};
@@ -23,58 +22,27 @@ use std::{collections::HashMap, sync::Arc};
 use crate::auth::EndpointAuth;
 use crate::{dataconnector::DataConnector, datafusion::DataFusion};
 use ::datafusion::error::DataFusionError;
-use ::datafusion::sql::parser::DFParser;
-use ::datafusion::sql::sqlparser::dialect::PostgreSqlDialect;
 use ::datafusion::sql::{sqlparser, TableReference};
-use ::opentelemetry::KeyValue;
-use accelerated_table::AcceleratedTable;
 use app::App;
 use builder::RuntimeBuilder;
-use cache::QueryResultsCacheProvider;
-use component::catalog::Catalog;
-use component::dataset::acceleration::RefreshMode;
-use component::dataset::{self, Dataset};
-use component::view::View;
 use config::Config;
-use dataconnector::localpod::{LocalPodConnector, LOCALPOD_DATACONNECTOR};
-use dataconnector::DataConnectorParams;
-use datafusion::SPICE_RUNTIME_SCHEMA;
 use datasets_health_monitor::DatasetsHealthMonitor;
-use embeddings::connector::EmbeddingConnector;
-use embeddings::task::TaskEmbed;
 use extension::ExtensionFactory;
-use federated_table::FederatedTable;
 use futures::future::join_all;
-use futures::{Future, StreamExt};
-use itertools::Itertools;
-use llms::chat::Chat;
-use llms::embeddings::Embed;
-use model::{
-    try_to_chat_model, try_to_embedding, EmbeddingModelStore, LLMModelStore,
-    ENABLE_MODEL_SUPPORT_MESSAGE,
-};
+use futures::Future;
+use model::{EmbeddingModelStore, LLMModelStore};
 use model_components::model::Model;
 pub use notify::Error as NotifyError;
 use secrecy::SecretString;
 use secrets::ParamStr;
 use snafu::prelude::*;
-use spice_metrics::get_metrics_table_reference;
-use spicepod::component::embeddings::Embeddings;
-use spicepod::component::model::{Model as SpicepodModel, ModelType};
-use spicepod::component::tool::Tool;
-use timing::TimeMeasurement;
 use tls::TlsConfig;
 use tokio::sync::oneshot::error::RecvError;
 use tokio::sync::RwLock;
 use tools::catalog::SpiceToolCatalog;
-use tools::factory::{self as tool_factory, default_available_catalogs};
 use tools::SpiceModelTool;
 use tools::Tooling;
-use topological_ordering::construct_effected_in_topological_order;
-use tracing_util::dataset_registered_trace;
-use util::fibonacci_backoff::FibonacciBackoffBuilder;
 pub use util::shutdown_signal;
-use util::{retry, RetryError};
 
 use crate::extension::Extension;
 pub mod accelerated_table;
@@ -93,6 +61,7 @@ pub mod extension;
 pub mod federated_table;
 mod flight;
 mod http;
+mod init;
 pub mod internal_table;
 mod metrics;
 mod metrics_server;
@@ -510,1400 +479,6 @@ impl Runtime {
 
         params_with_secrets
     }
-
-    async fn start_extensions(&self) {
-        let mut extensions = self.extensions.write().await;
-        for (name, extension) in extensions.iter_mut() {
-            if let Err(err) = extension.on_start(self).await {
-                tracing::warn!("Failed to start extension {name}: {err}");
-            }
-        }
-    }
-
-    fn datasets_iter(app: &Arc<App>) -> impl Iterator<Item = Result<Dataset>> + '_ {
-        app.datasets
-            .clone()
-            .into_iter()
-            .map(Dataset::try_from)
-            .map(move |ds| ds.map(|ds| Dataset::with_app(ds, Arc::clone(app))))
-    }
-
-    fn catalogs_iter<'a>(app: &Arc<App>) -> impl Iterator<Item = Result<Catalog>> + 'a {
-        app.catalogs.clone().into_iter().map(Catalog::try_from)
-    }
-
-    /// Returns a list of valid datasets from the given App, skipping any that fail to parse and logging an error for them.
-    fn get_valid_datasets(app: &Arc<App>, log_errors: LogErrors) -> Vec<Arc<Dataset>> {
-        Self::datasets_iter(app)
-            .zip(&app.datasets)
-            .filter_map(|(ds, spicepod_ds)| match ds {
-                Ok(ds) => Some(Arc::new(ds)),
-                Err(e) => {
-                    if log_errors.0 {
-                        metrics::datasets::LOAD_ERROR.add(1, &[]);
-                        tracing::error!(dataset = &spicepod_ds.name, "{e}");
-                    }
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Returns a list of valid catalogs from the given App, skipping any that fail to parse and logging an error for them.
-    fn get_valid_catalogs(app: &Arc<App>, log_errors: LogErrors) -> Vec<Catalog> {
-        Self::catalogs_iter(app)
-            .zip(&app.catalogs)
-            .filter_map(|(catalog, spicepod_catalog)| match catalog {
-                Ok(catalog) => Some(catalog),
-                Err(e) => {
-                    if log_errors.0 {
-                        metrics::catalogs::LOAD_ERROR.add(1, &[]);
-                        tracing::error!(catalog = &spicepod_catalog.name, "{e}");
-                    }
-                    None
-                }
-            })
-            .collect()
-    }
-
-    /// Returns a list of valid views from the given App, skipping any that fail to parse and logging an error for them.
-    fn get_valid_views(app: &Arc<App>, log_errors: LogErrors) -> Vec<View> {
-        let datasets = Self::get_valid_datasets(app, log_errors)
-            .iter()
-            .map(|ds| ds.name.clone())
-            .collect::<HashSet<_>>();
-
-        app.views
-            .iter()
-            .cloned()
-            .map(View::try_from)
-            .zip(&app.views)
-            .filter_map(|(view, spicepod_view)| match view {
-                Ok(view) => {
-                    // only load this view if the name isn't used by an existing dataset
-                    if datasets.contains(&view.name) {
-                        if log_errors.0 {
-                            metrics::views::LOAD_ERROR.add(1, &[]);
-                            tracing::error!(
-                                view = &spicepod_view.name,
-                                "View name is already in use by a dataset."
-                            );
-                        }
-                        None
-                    } else {
-                        Some(view)
-                    }
-                }
-                Err(e) => {
-                    if log_errors.0 {
-                        metrics::views::LOAD_ERROR.add(1, &[]);
-                        tracing::error!(view = &spicepod_view.name, "{e}");
-                    }
-                    None
-                }
-            })
-            .collect()
-    }
-
-    async fn load_catalogs(&self) {
-        let app_lock = self.app.read().await;
-        let Some(app) = app_lock.as_ref() else {
-            return;
-        };
-
-        let valid_catalogs = Self::get_valid_catalogs(app, LogErrors(true));
-        let mut futures = vec![];
-        for catalog in &valid_catalogs {
-            self.status
-                .update_catalog(&catalog.name, status::ComponentStatus::Initializing);
-            futures.push(self.load_catalog(catalog));
-        }
-
-        let _ = join_all(futures).await;
-    }
-
-    /// Returns a list of valid datasets from the given App, skipping any that fail to parse and logging an error for them.
-    async fn get_initialized_datasets(app: &Arc<App>, log_errors: LogErrors) -> Vec<Arc<Dataset>> {
-        let valid_datasets = Self::get_valid_datasets(app, log_errors);
-        futures::stream::iter(valid_datasets)
-            .filter_map(|ds| async move {
-                match (ds.is_accelerated(), ds.is_accelerator_initialized().await) {
-                    (true, true) | (false, _) => Some(Arc::clone(&ds)),
-                    (true, false) => {
-                        if log_errors.0 {
-                            metrics::datasets::LOAD_ERROR.add(1, &[]);
-                            tracing::error!(
-                                dataset = &ds.name.to_string(),
-                                "Dataset is accelerated but the accelerator failed to initialize."
-                            );
-                        }
-                        None
-                    }
-                }
-            })
-            .collect()
-            .await
-    }
-
-    /// Initialize datasets configured with accelerators before registering the datasets.
-    /// This ensures that the required resources for acceleration are available before registration,
-    /// which is important for acceleration federation for some acceleration engines (e.g. `SQLite`).
-    async fn initialize_accelerators(&self, datasets: &[Arc<Dataset>]) -> Vec<Arc<Dataset>> {
-        let spaced_tracer = Arc::clone(&self.spaced_tracer);
-
-        let mut initialized_datasets = vec![];
-        for ds in datasets {
-            if let Some(acceleration) = &ds.acceleration {
-                let accelerator = match dataaccelerator::get_accelerator_engine(acceleration.engine)
-                    .await
-                    .context(AcceleratorEngineNotAvailableSnafu {
-                        name: acceleration.engine.to_string(),
-                    }) {
-                    Ok(accelerator) => accelerator,
-                    Err(err) => {
-                        let ds_name = &ds.name;
-                        self.status
-                            .update_dataset(ds_name, status::ComponentStatus::Error);
-                        metrics::datasets::LOAD_ERROR.add(1, &[]);
-                        warn_spaced!(spaced_tracer, "{} {err}", ds_name.table());
-                        continue;
-                    }
-                };
-
-                match accelerator
-                    .init(ds)
-                    .await
-                    .context(AcceleratorInitializationFailedSnafu {
-                        name: acceleration.engine.to_string(),
-                    }) {
-                    Ok(()) => {
-                        initialized_datasets.push(Arc::clone(ds));
-                    }
-                    Err(err) => {
-                        let ds_name = &ds.name;
-                        self.status
-                            .update_dataset(ds_name, status::ComponentStatus::Error);
-                        metrics::datasets::LOAD_ERROR.add(1, &[]);
-                        warn_spaced!(spaced_tracer, "{} {err}", ds_name.table());
-                    }
-                }
-            } else {
-                initialized_datasets.push(Arc::clone(ds)); // non-accelerated datasets are always successfully initialized
-            }
-        }
-
-        initialized_datasets
-    }
-
-    async fn load_datasets(&self) {
-        let app_lock = self.app.read().await;
-        let Some(app) = app_lock.as_ref() else {
-            return;
-        };
-
-        let valid_datasets = Self::get_valid_datasets(app, LogErrors(true));
-        let initialized_datasets = self.initialize_accelerators(&valid_datasets).await;
-
-        // Create a map of dataset names to their futures
-        let mut dataset_futures = HashMap::new();
-        let mut localpod_datasets = Vec::new();
-
-        // First create futures for non-localpod datasets
-        for ds in initialized_datasets {
-            if ds.source() == LOCALPOD_DATACONNECTOR {
-                localpod_datasets.push(ds);
-                continue;
-            }
-
-            self.status
-                .update_dataset(&ds.name, status::ComponentStatus::Initializing);
-            let future: Pin<Box<dyn Future<Output = ()> + Send>> =
-                Box::pin(self.load_dataset(ds.clone()));
-            dataset_futures.insert(ds.name.clone(), future);
-        }
-
-        // For each localpod dataset, chain it after its parent's future
-        for ds in localpod_datasets {
-            self.status
-                .update_dataset(&ds.name, status::ComponentStatus::Initializing);
-
-            // Get the parent dataset path from the localpod dataset
-            let path = ds.path();
-            let path_table_ref = TableReference::parse_str(&path);
-
-            // Find and remove the parent dataset's future
-            match dataset_futures.remove(&path_table_ref) {
-                Some(parent_future) => {
-                    let ds_clone = Arc::clone(&ds);
-
-                    // Chain the localpod dataset load after its parent
-                    let chained_future = Box::pin(async move {
-                        parent_future.await;
-                        self.load_dataset(ds_clone).await;
-                    })
-                        as Pin<Box<dyn Future<Output = ()> + Send>>;
-
-                    // Replace parent future with the chained future
-                    dataset_futures.insert(ds.name.clone(), chained_future);
-                }
-                None => {
-                    // If parent doesn't exist, create a future that will fail
-                    tracing::error!(
-                        "Parent dataset {} not found for localpod dataset {}",
-                        path_table_ref,
-                        ds.name
-                    );
-                    self.status
-                        .update_dataset(&ds.name, status::ComponentStatus::Error);
-                    continue;
-                }
-            };
-        }
-
-        // Execute all futures
-        if let Some(parallel_num) = app.runtime.num_of_parallel_loading_at_start_up {
-            let stream =
-                futures::stream::iter(dataset_futures.into_values()).buffer_unordered(parallel_num);
-            let _ = stream.collect::<Vec<_>>().await;
-        } else {
-            let _ = join_all(dataset_futures.into_values()).await;
-        }
-
-        // After all datasets have loaded, load the views.
-        self.load_views(app);
-    }
-
-    fn load_views(&self, app: &Arc<App>) {
-        let views: Vec<View> = Self::get_valid_views(app, LogErrors(true));
-
-        for view in &views {
-            if let Err(e) = self.load_view(view) {
-                tracing::error!("Unable to load view: {e}");
-            };
-        }
-    }
-
-    async fn load_catalog(&self, catalog: &Catalog) {
-        let spaced_tracer = Arc::clone(&self.spaced_tracer);
-
-        let retry_strategy = FibonacciBackoffBuilder::new().max_retries(None).build();
-
-        let _ = retry(retry_strategy, || async {
-            let connector = match self.load_catalog_connector(catalog).await {
-                Ok(connector) => connector,
-                Err(err) => {
-                    let catalog_name = &catalog.name;
-                    self.status
-                        .update_catalog(catalog_name, status::ComponentStatus::Error);
-                    metrics::catalogs::LOAD_ERROR.add(1, &[]);
-                    warn_spaced!(spaced_tracer, "{} {err}", catalog_name);
-                    return Err(RetryError::transient(err));
-                }
-            };
-
-            if let Err(err) = self.register_catalog(catalog, connector).await {
-                tracing::error!("Unable to register catalog {}: {err}", &catalog.name);
-                return Err(RetryError::transient(err));
-            };
-
-            self.status
-                .update_catalog(&catalog.name, status::ComponentStatus::Ready);
-
-            Ok(())
-        })
-        .await;
-    }
-
-    /// Caller must set `status::update_dataset(...` before calling `load_dataset`. This function will set error/ready statuses appropriately.
-    async fn load_dataset(&self, ds: Arc<Dataset>) {
-        let spaced_tracer = Arc::clone(&self.spaced_tracer);
-
-        let retry_strategy = FibonacciBackoffBuilder::new().max_retries(None).build();
-
-        let _ = retry(retry_strategy, || async {
-            let connector = match self.load_dataset_connector(Arc::clone(&ds)).await {
-                Ok(connector) => connector,
-                Err(err) => {
-                    let ds_name = &ds.name;
-                    self.status
-                        .update_dataset(ds_name, status::ComponentStatus::Error);
-                    metrics::datasets::LOAD_ERROR.add(1, &[]);
-                    warn_spaced!(
-                        spaced_tracer,
-                        "Error initializing dataset {}. {err}",
-                        ds_name.table()
-                    );
-                    return Err(RetryError::transient(err));
-                }
-            };
-
-            if let Err(err) = self
-                .register_loaded_dataset(Arc::clone(&ds), connector, None)
-                .await
-            {
-                return Err(RetryError::transient(err));
-            };
-
-            Ok(())
-        })
-        .await;
-    }
-
-    fn load_view(&self, view: &View) -> Result<()> {
-        let df = Arc::clone(&self.df);
-        df.register_view(view.name.clone(), view.sql.clone())
-            .context(UnableToAttachViewSnafu)
-            .inspect_err(|_| {
-                self.status
-                    .update_view(&view.name, status::ComponentStatus::Error);
-            })?;
-
-        self.status
-            .update_view(&view.name, status::ComponentStatus::Ready);
-        Ok(())
-    }
-
-    fn remove_view(&self, view: &View) {
-        if self.df.table_exists(view.name.clone()) {
-            if let Err(e) = self.df.remove_view(&view.name) {
-                tracing::warn!("Unable to unload view {}: {}", &view.name, e);
-                return;
-            }
-        }
-        tracing::info!("Unloaded view {}", &view.name);
-    }
-
-    fn update_view(&self, view: &View) {
-        self.status
-            .update_view(&view.name, status::ComponentStatus::Refreshing);
-        self.remove_view(view);
-        let _ = self.load_view(view);
-    }
-
-    async fn load_catalog_connector(&self, catalog: &Catalog) -> Result<Arc<dyn DataConnector>> {
-        let spaced_tracer = Arc::clone(&self.spaced_tracer);
-        let catalog = catalog.clone();
-
-        let source = catalog.provider;
-        let params = catalog.params.clone();
-        let params = DataConnectorParams::from_params(self, &source, params)
-            .await
-            .context(UnableToInitializeDataConnectorSnafu)?;
-        let data_connector: Arc<dyn DataConnector> =
-            match self.get_dataconnector_from_source(&source, params).await {
-                Ok(data_connector) => data_connector,
-                Err(err) => {
-                    let catalog_name = &catalog.name;
-                    self.status
-                        .update_catalog(catalog_name, status::ComponentStatus::Error);
-                    metrics::catalogs::LOAD_ERROR.add(1, &[]);
-                    warn_spaced!(spaced_tracer, "{} {err}", catalog_name);
-                    return UnableToLoadDatasetConnectorSnafu {
-                        dataset: catalog_name.clone(),
-                    }
-                    .fail();
-                }
-            };
-
-        Ok(data_connector)
-    }
-
-    async fn load_dataset_connector(&self, ds: Arc<Dataset>) -> Result<Arc<dyn DataConnector>> {
-        let spaced_tracer = Arc::clone(&self.spaced_tracer);
-
-        let source = ds.source();
-        let params = DataConnectorParams::from_dataset(self, Arc::clone(&ds))
-            .await
-            .context(UnableToInitializeDataConnectorSnafu)?;
-
-        let data_connector: Arc<dyn DataConnector> =
-            match self.get_dataconnector_from_source(&source, params).await {
-                Ok(data_connector) => data_connector,
-                Err(err) => {
-                    let ds_name = &ds.name;
-                    self.status
-                        .update_dataset(ds_name, status::ComponentStatus::Error);
-                    metrics::datasets::LOAD_ERROR.add(1, &[]);
-                    warn_spaced!(
-                        spaced_tracer,
-                        "Error initializing dataset {}. {err}",
-                        ds_name.table()
-                    );
-                    return UnableToLoadDatasetConnectorSnafu {
-                        dataset: ds.name.clone(),
-                    }
-                    .fail();
-                }
-            };
-
-        Ok(data_connector)
-    }
-
-    async fn register_catalog(
-        &self,
-        catalog: &Catalog,
-        data_connector: Arc<dyn DataConnector>,
-    ) -> Result<()> {
-        tracing::info!(
-            "Registering catalog '{}' for {}",
-            &catalog.name,
-            &catalog.provider
-        );
-        let catalog_provider = data_connector
-            .catalog_provider(self, catalog)
-            .await
-            .context(DataConnectorDoesntSupportCatalogsSnafu {
-                dataconnector: catalog.provider.clone(),
-            })?
-            .boxed()
-            .context(UnableToLoadCatalogConnectorSnafu {
-                catalog: catalog.name.clone(),
-            })?;
-        let num_schemas = catalog_provider
-            .schema_names()
-            .iter()
-            .fold(0, |acc, schema| {
-                acc + catalog_provider
-                    .schema(schema)
-                    .map_or(0, |s| i32::from(!s.table_names().is_empty()))
-            });
-        let num_tables = catalog_provider
-            .schema_names()
-            .iter()
-            .fold(0, |acc, schema| {
-                acc + catalog_provider
-                    .schema(schema)
-                    .map_or(0, |s| s.table_names().len())
-            });
-
-        self.df
-            .register_catalog(&catalog.name, catalog_provider)
-            .boxed()
-            .context(UnableToLoadCatalogConnectorSnafu {
-                catalog: catalog.name.clone(),
-            })?;
-
-        tracing::info!(
-            "Registered catalog '{}' with {num_schemas} schema{} and {num_tables} table{}",
-            &catalog.name,
-            if num_schemas == 1 { "" } else { "s" },
-            if num_tables == 1 { "" } else { "s" },
-        );
-
-        Ok(())
-    }
-
-    async fn register_loaded_dataset(
-        &self,
-        ds: Arc<Dataset>,
-        data_connector: Arc<dyn DataConnector>,
-        accelerated_table: Option<AcceleratedTable>,
-    ) -> Result<()> {
-        let source = ds.source();
-        let spaced_tracer = Arc::clone(&self.spaced_tracer);
-        if let Some(acceleration) = &ds.acceleration {
-            if data_connector.resolve_refresh_mode(acceleration.refresh_mode)
-                == RefreshMode::Changes
-                && !data_connector.supports_changes_stream()
-            {
-                let err = AcceleratedTableInvalidChangesSnafu {
-                    dataset_name: ds.name.to_string(),
-                }
-                .build();
-                warn_spaced!(spaced_tracer, "{}{err}", "");
-                return Err(err);
-            }
-        }
-
-        // Only wrap data connector when necessary.
-        let connector = if ds.has_embeddings() {
-            let connector = EmbeddingConnector::new(data_connector, Arc::clone(&self.embeds));
-            Arc::new(connector) as Arc<dyn DataConnector>
-        } else {
-            data_connector
-        };
-
-        // Test dataset connectivity by attempting to get a read provider.
-        let federated_table = match connector.read_provider(&ds).await {
-            Ok(provider) => FederatedTable::new(provider),
-            Err(err) => {
-                // We couldn't connect to the federated table. If the dataset has an existing
-                // accelerated table, we can defer the federated table creation.
-                if let Some(federated_table) =
-                    FederatedTable::new_deferred(Arc::clone(&ds), Arc::clone(&connector)).await
-                {
-                    tracing::warn!(
-                        "Connection to source failed for {}. Data from acceleration is available.\n\n{err}",
-                        ds.name
-                    );
-                    federated_table
-                } else {
-                    self.status
-                        .update_dataset(&ds.name, status::ComponentStatus::Error);
-                    metrics::datasets::LOAD_ERROR.add(1, &[]);
-                    warn_spaced!(spaced_tracer, "{}{err}", "");
-                    return UnableToLoadDatasetConnectorSnafu {
-                        dataset: ds.name.clone(),
-                    }
-                    .fail();
-                }
-            }
-        };
-
-        match self
-            .register_dataset(
-                Arc::clone(&ds),
-                RegisterDatasetContext {
-                    data_connector: Arc::clone(&connector),
-                    federated_read_table: federated_table,
-                    source,
-                    accelerated_table,
-                },
-            )
-            .await
-        {
-            Ok(()) => {
-                tracing::info!(
-                    "{}",
-                    dataset_registered_trace(&connector, &ds, self.df.cache_provider().is_some())
-                );
-                if let Some(datasets_health_monitor) = &self.datasets_health_monitor {
-                    if let Err(err) = datasets_health_monitor.register_dataset(&ds).await {
-                        tracing::warn!(
-                            "Unable to add dataset {} for availability monitoring: {err}",
-                            &ds.name
-                        );
-                    };
-                }
-                let engine = ds.acceleration.as_ref().map_or_else(
-                    || "None".to_string(),
-                    |acc| {
-                        if acc.enabled {
-                            acc.engine.to_string()
-                        } else {
-                            "None".to_string()
-                        }
-                    },
-                );
-                metrics::datasets::COUNT.add(1, &[KeyValue::new("engine", engine)]);
-
-                Ok(())
-            }
-            Err(err) => {
-                self.status
-                    .update_dataset(&ds.name, status::ComponentStatus::Error);
-                metrics::datasets::LOAD_ERROR.add(1, &[]);
-                if let Error::UnableToAttachDataConnector {
-                    source: datafusion::Error::RefreshSql { source },
-                    data_connector: _,
-                } = &err
-                {
-                    tracing::error!("{source}");
-                }
-                warn_spaced!(spaced_tracer, "{}{err}", "");
-
-                Err(err)
-            }
-        }
-    }
-
-    async fn remove_dataset(&self, ds: &Dataset) {
-        if self.df.table_exists(ds.name.clone()) {
-            if let Some(datasets_health_monitor) = &self.datasets_health_monitor {
-                datasets_health_monitor
-                    .deregister_dataset(&ds.name.to_string())
-                    .await;
-            }
-
-            if let Err(e) = self.df.remove_table(&ds.name).await {
-                tracing::warn!("Unable to unload dataset {}: {}", &ds.name, e);
-                return;
-            }
-        }
-
-        tracing::info!("Unloaded dataset {}", &ds.name);
-        let engine = ds.acceleration.as_ref().map_or_else(
-            || "None".to_string(),
-            |acc| {
-                if acc.enabled {
-                    acc.engine.to_string()
-                } else {
-                    "None".to_string()
-                }
-            },
-        );
-        metrics::datasets::COUNT.add(-1, &[KeyValue::new("engine", engine)]);
-    }
-
-    async fn update_dataset(&self, ds: Arc<Dataset>) {
-        self.status
-            .update_dataset(&ds.name, status::ComponentStatus::Refreshing);
-        match self.load_dataset_connector(Arc::clone(&ds)).await {
-            Ok(connector) => {
-                // File accelerated datasets don't support hot reload.
-                if Self::accelerated_dataset_supports_hot_reload(&ds, &*connector) {
-                    tracing::info!("Updating accelerated dataset {}...", &ds.name);
-                    if let Ok(()) = &self
-                        .reload_accelerated_dataset(Arc::clone(&ds), Arc::clone(&connector))
-                        .await
-                    {
-                        self.status
-                            .update_dataset(&ds.name, status::ComponentStatus::Ready);
-                        return;
-                    }
-                    tracing::debug!("Failed to create accelerated table for dataset {}, falling back to full dataset reload", ds.name);
-                }
-
-                self.remove_dataset(&ds).await;
-
-                // Initialize file mode accelerator when reloading with file mode acceleration
-                // Fail when there's no successfully initiated dataset
-                if ds.is_file_accelerated() {
-                    let datasets = self.initialize_accelerators(&[Arc::clone(&ds)]).await;
-                    if datasets.is_empty() {
-                        return;
-                    }
-                }
-
-                if self
-                    .register_loaded_dataset(Arc::clone(&ds), Arc::clone(&connector), None)
-                    .await
-                    .is_err()
-                {
-                    self.status
-                        .update_dataset(&ds.name, status::ComponentStatus::Error);
-                }
-            }
-            Err(e) => {
-                tracing::error!("Unable to update dataset {}: {e}", ds.name);
-                self.status
-                    .update_dataset(&ds.name, status::ComponentStatus::Error);
-            }
-        }
-    }
-
-    fn accelerated_dataset_supports_hot_reload(
-        ds: &Dataset,
-        connector: &dyn DataConnector,
-    ) -> bool {
-        let Some(acceleration) = &ds.acceleration else {
-            return false;
-        };
-
-        if !acceleration.enabled {
-            return false;
-        }
-
-        // Datasets that configure changes and are file-accelerated automatically keep track of changes that survive restarts.
-        // Thus we don't need to "hot reload" them to try to keep their data intact.
-        if connector.supports_changes_stream()
-            && ds.is_file_accelerated()
-            && connector.resolve_refresh_mode(acceleration.refresh_mode) == RefreshMode::Changes
-        {
-            return false;
-        }
-
-        // File accelerated datasets don't support hot reload.
-        if ds.is_file_accelerated() {
-            return false;
-        }
-
-        true
-    }
-
-    async fn reload_accelerated_dataset(
-        &self,
-        ds: Arc<Dataset>,
-        connector: Arc<dyn DataConnector>,
-    ) -> Result<()> {
-        let read_table = connector.read_provider(&ds).await.map_err(|_| {
-            UnableToLoadDatasetConnectorSnafu {
-                dataset: ds.name.clone(),
-            }
-            .build()
-        })?;
-        let federated_table = FederatedTable::new(read_table);
-
-        // create new accelerated table for updated data connector
-        let (accelerated_table, is_ready) = self
-            .df
-            .create_accelerated_table(&ds, Arc::clone(&connector), federated_table, self.secrets())
-            .await
-            .context(UnableToCreateAcceleratedTableSnafu {
-                dataset: ds.name.clone(),
-            })?;
-
-        // wait for accelerated table to be ready
-        is_ready
-            .await
-            .context(UnableToReceiveAcceleratedTableStatusSnafu)?;
-
-        tracing::debug!("Accelerated table for dataset {} is ready", ds.name);
-
-        self.register_loaded_dataset(ds, Arc::clone(&connector), Some(accelerated_table))
-            .await?;
-
-        Ok(())
-    }
-
-    async fn get_dataconnector_from_source(
-        &self,
-        source: &str,
-        params: DataConnectorParams,
-    ) -> Result<Arc<dyn DataConnector>> {
-        // Unlike most other data connectors, the localpod connector needs a reference to the current DataFusion instance.
-        if source == LOCALPOD_DATACONNECTOR {
-            return Ok(Arc::new(LocalPodConnector::new(Arc::clone(&self.df))));
-        }
-
-        match dataconnector::create_new_connector(source, params).await {
-            Some(dc) => dc.context(UnableToInitializeDataConnectorSnafu {}),
-            None => UnknownDataConnectorSnafu {
-                data_connector: source,
-            }
-            .fail(),
-        }
-    }
-
-    async fn register_dataset(
-        &self,
-        ds: Arc<Dataset>,
-        register_dataset_ctx: RegisterDatasetContext,
-    ) -> Result<()> {
-        let RegisterDatasetContext {
-            data_connector,
-            federated_read_table,
-            source,
-            accelerated_table,
-        } = register_dataset_ctx;
-
-        let replicate = ds.replication.as_ref().map_or(false, |r| r.enabled);
-
-        // FEDERATED TABLE
-        if !ds.is_accelerated() {
-            let ds_name: TableReference = ds.name.clone();
-            self.df
-                .register_table(
-                    ds,
-                    datafusion::Table::Federated {
-                        data_connector,
-                        federated_read_table,
-                    },
-                )
-                .await
-                .context(UnableToAttachDataConnectorSnafu {
-                    data_connector: source,
-                })?;
-
-            self.status
-                .update_dataset(&ds_name, status::ComponentStatus::Ready);
-            return Ok(());
-        }
-
-        // ACCELERATED TABLE
-        let acceleration_settings =
-            ds.acceleration
-                .as_ref()
-                .ok_or_else(|| Error::ExpectedAccelerationSettings {
-                    name: ds.name.to_string(),
-                })?;
-        let accelerator_engine = acceleration_settings.engine;
-
-        if ds.mode() == dataset::Mode::ReadWrite && !replicate {
-            AcceleratedReadWriteTableWithoutReplicationSnafu.fail()?;
-        }
-
-        dataaccelerator::get_accelerator_engine(accelerator_engine)
-            .await
-            .context(AcceleratorEngineNotAvailableSnafu {
-                name: accelerator_engine.to_string(),
-            })?;
-
-        // The accelerated refresh task will set the dataset status to `Ready` once it finishes loading.
-        self.status
-            .update_dataset(&ds.name, status::ComponentStatus::Refreshing);
-        self.df
-            .register_table(
-                ds,
-                datafusion::Table::Accelerated {
-                    source: data_connector,
-                    federated_read_table,
-                    accelerated_table,
-                    secrets: self.secrets(),
-                },
-            )
-            .await
-            .context(UnableToAttachDataConnectorSnafu {
-                data_connector: source,
-            })
-    }
-
-    /// Loads a specific LLM from the spicepod. If an error occurs, no retry attempt is made.
-    async fn load_llm(
-        &self,
-        m: SpicepodModel,
-        params: HashMap<String, SecretString>,
-    ) -> Result<Box<dyn Chat>> {
-        let l = try_to_chat_model(&m, &params, Arc::new(self.clone()))
-            .await
-            .boxed()
-            .context(UnableToInitializeLlmSnafu)?;
-
-        l.health()
-            .await
-            .boxed()
-            .context(UnableToInitializeLlmSnafu)?;
-        Ok(l)
-    }
-
-    #[allow(dead_code)]
-    /// Loads a specific Embedding model from the spicepod. If an error occurs, no retry attempt is made.
-    async fn load_embedding(&self, in_embed: &Embeddings) -> Result<TaskEmbed> {
-        let params_with_secrets = self.get_params_with_secrets(&in_embed.params).await;
-
-        let l = try_to_embedding(in_embed, &params_with_secrets)
-            .boxed()
-            .context(UnableToInitializeEmbeddingModelSnafu)?;
-        l.health()
-            .await
-            .boxed()
-            .context(UnableToInitializeEmbeddingModelSnafu)?;
-
-        TaskEmbed::new(l)
-            .await
-            .boxed()
-            .context(UnableToInitializeEmbeddingModelSnafu)
-    }
-
-    #[allow(dead_code)]
-    async fn load_embeddings(&self) {
-        let app_opt = self.app.read().await;
-
-        if !cfg!(feature = "models") && app_opt.as_ref().is_some_and(|s| !s.embeddings.is_empty()) {
-            tracing::error!("Cannot load embedding models without the 'models' feature enabled. {ENABLE_MODEL_SUPPORT_MESSAGE}");
-            return;
-        };
-
-        if let Some(app) = app_opt.as_ref() {
-            for in_embed in &app.embeddings {
-                self.status
-                    .update_embedding(&in_embed.name, status::ComponentStatus::Initializing);
-                match self.load_embedding(in_embed).await {
-                    Ok(e) => {
-                        let mut embeds_map = self.embeds.write().await;
-
-                        embeds_map.insert(in_embed.name.clone(), Box::new(e) as Box<dyn Embed>);
-
-                        tracing::info!("Embedding [{}] ready to embed", in_embed.name);
-                        metrics::embeddings::COUNT.add(
-                            1,
-                            &[
-                                KeyValue::new("embeddings", in_embed.name.clone()),
-                                KeyValue::new(
-                                    "source",
-                                    in_embed
-                                        .get_prefix()
-                                        .map(|x| x.to_string())
-                                        .unwrap_or_default(),
-                                ),
-                            ],
-                        );
-                        self.status
-                            .update_embedding(&in_embed.name, status::ComponentStatus::Ready);
-                    }
-                    Err(e) => {
-                        metrics::embeddings::LOAD_ERROR.add(1, &[]);
-                        self.status
-                            .update_embedding(&in_embed.name, status::ComponentStatus::Error);
-                        tracing::warn!(
-                            "Unable to load embedding from spicepod {}, error: {}",
-                            in_embed.name,
-                            e,
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    async fn load_models(&self) {
-        let app_lock = self.app.read().await;
-
-        if !cfg!(feature = "models") && app_lock.as_ref().is_some_and(|s| !s.models.is_empty()) {
-            tracing::error!("Cannot load models without the 'models' feature enabled. {ENABLE_MODEL_SUPPORT_MESSAGE}");
-            return;
-        }
-
-        // Load tools before loading models.
-        self.load_tools().await;
-
-        if let Some(app) = app_lock.as_ref() {
-            for model in &app.models {
-                self.status
-                    .update_model(&model.name, status::ComponentStatus::Initializing);
-                self.load_model(model).await;
-            }
-        }
-    }
-
-    #[allow(clippy::implicit_hasher)]
-    async fn load_tools(&self) {
-        let app_lock = self.app.read().await;
-        if let Some(app) = app_lock.as_ref() {
-            for tool in &app.tools {
-                self.load_tool(tool).await;
-            }
-        }
-
-        // Load all built-in tools, regardless if they are in the spicepod.
-        // This will enable loading each tool in the catalog, and the catalog as a whole. E.g:
-        //   `tools: models, builtin`
-        //   `tools: sql, load_memory`
-        for ctlg in default_available_catalogs() {
-            self.insert_tool_catalog(&ctlg).await;
-            for tool in ctlg.all().await {
-                self.insert_tool(&tool).await;
-            }
-        }
-    }
-
-    async fn insert_tool_catalog(&self, t: &Arc<dyn SpiceToolCatalog>) {
-        let name = t.name().to_string();
-        let mut tools_map = self.tools.write().await;
-
-        tools_map.insert(name.clone(), Arc::clone(t).into());
-        tracing::debug!("Tool catalog {} ready to use", name.clone());
-        metrics::tools::COUNT.add(1, &[KeyValue::new("tool_catalog", name.clone())]);
-        self.status
-            .update_tool_catalog(&name, status::ComponentStatus::Ready);
-    }
-
-    async fn insert_tool(&self, t: &Arc<dyn SpiceModelTool>) {
-        let name = t.name().to_string();
-        let mut tools_map = self.tools.write().await;
-
-        tools_map.insert(name.clone(), Arc::clone(t).into());
-        tracing::debug!("Tool {} ready to use", name.clone());
-        metrics::tools::COUNT.add(1, &[KeyValue::new("tool", name.clone())]);
-        self.status
-            .update_tool(&name, status::ComponentStatus::Ready);
-    }
-
-    async fn load_tool(&self, tool: &Tool) {
-        self.status
-            .update_tool(&tool.name, status::ComponentStatus::Initializing);
-        let params_with_secrets: HashMap<String, SecretString> =
-            self.get_params_with_secrets(&tool.params).await;
-
-        match tool_factory::forge(tool, params_with_secrets)
-            .await
-            .context(UnableToInitializeLlmToolSnafu)
-        {
-            Ok(t) => self.insert_tool(&t).await,
-            Err(e) => {
-                metrics::tools::LOAD_ERROR.add(1, &[]);
-                self.status
-                    .update_tool(&tool.name, status::ComponentStatus::Error);
-                tracing::warn!(
-                    "Unable to load tool from spicepod {}, error: {}",
-                    tool.name,
-                    e,
-                );
-            }
-        }
-    }
-
-    // Caller must set `status::update_model(...` before calling `load_model`. This function will set error/ready statues appropriately.`
-    async fn load_model(&self, m: &SpicepodModel) {
-        let source = m.get_source();
-        let source_str = source.clone().map(|s| s.to_string()).unwrap_or_default();
-        let model = m.clone();
-        let _guard = TimeMeasurement::new(
-            &metrics::models::LOAD_DURATION_MS,
-            &[
-                KeyValue::new("model", m.name.clone()),
-                KeyValue::new("source", source_str.clone()),
-            ],
-        );
-
-        tracing::info!("Loading model [{}] from {}...", m.name, m.from);
-
-        // TODO: Have downstream code using model parameters to accept `Hashmap<String, Value>`.
-        // This will require handling secrets with `Value` type.
-        let p = m
-            .params
-            .clone()
-            .iter()
-            .map(|(k, v)| {
-                let k = k.clone();
-                match v.as_str() {
-                    Some(s) => (k, s.to_string()),
-                    None => (k, v.to_string()),
-                }
-            })
-            .collect::<HashMap<_, _>>();
-        let params = self.get_params_with_secrets(&p).await;
-
-        let model_type = m.model_type();
-        tracing::trace!("Model type for {} is {:#?}", m.name, model_type.clone());
-        let result: Result<(), String> = match model_type {
-            Some(ModelType::Llm) => match self.load_llm(m.clone(), params).await {
-                Ok(l) => {
-                    let mut llm_map = self.llms.write().await;
-                    llm_map.insert(m.name.clone(), l);
-                    Ok(())
-                }
-                Err(e) => Err(format!(
-                    "Unable to load LLM from spicepod {}, error: {}",
-                    m.name, e,
-                )),
-            },
-            Some(ModelType::Ml) => match Model::load(m.clone(), params).await {
-                Ok(in_m) => {
-                    let mut model_map = self.models.write().await;
-                    model_map.insert(m.name.clone(), in_m);
-                    Ok(())
-                }
-                Err(e) => Err(format!(
-                    "Unable to load runnable model from spicepod {}, error: {}",
-                    m.name, e,
-                )),
-            },
-            None => Err(format!(
-                "Unable to load model {} from spicepod. Unable to determine model type.",
-                m.name,
-            )),
-        };
-        match result {
-            Ok(()) => {
-                tracing::info!("Model [{}] deployed, ready for inferencing", m.name);
-                metrics::models::COUNT.add(
-                    1,
-                    &[
-                        KeyValue::new("model", m.name.clone()),
-                        KeyValue::new("source", source_str),
-                    ],
-                );
-                self.status
-                    .update_model(&model.name, status::ComponentStatus::Ready);
-            }
-            Err(e) => {
-                metrics::models::LOAD_ERROR.add(1, &[]);
-                self.status
-                    .update_model(&model.name, status::ComponentStatus::Error);
-                tracing::warn!(e);
-            }
-        }
-    }
-
-    async fn remove_model(&self, m: &SpicepodModel) {
-        match m.model_type() {
-            Some(ModelType::Ml) => {
-                let mut ml_map = self.models.write().await;
-                ml_map.remove(&m.name);
-            }
-            Some(ModelType::Llm) => {
-                let mut llm_map = self.llms.write().await;
-                llm_map.remove(&m.name);
-            }
-            None => return,
-        };
-
-        tracing::info!("Model [{}] has been unloaded", m.name);
-        let source_str = m.get_source().map(|s| s.to_string()).unwrap_or_default();
-        metrics::models::COUNT.add(
-            -1,
-            &[
-                KeyValue::new("model", m.name.clone()),
-                KeyValue::new("source", source_str),
-            ],
-        );
-    }
-
-    async fn update_model(&self, m: &SpicepodModel) {
-        self.status
-            .update_model(&m.name, status::ComponentStatus::Refreshing);
-        self.remove_model(m).await;
-        self.load_model(m).await;
-    }
-
-    async fn register_metrics_table(&self, metrics_enabled: bool) -> Result<()> {
-        if metrics_enabled {
-            let table_reference = get_metrics_table_reference();
-            let metrics_table = self.df.get_table(&table_reference).await;
-
-            if metrics_table.is_none() {
-                tracing::debug!("Registering local metrics table");
-                spice_metrics::register_metrics_table(&self.df)
-                    .await
-                    .context(UnableToStartLocalMetricsSnafu)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    async fn apply_model_diff(&self, current_app: &Arc<App>, new_app: &Arc<App>) {
-        for model in &new_app.models {
-            if let Some(current_model) = current_app.models.iter().find(|m| m.name == model.name) {
-                if current_model != model {
-                    self.update_model(model).await;
-                }
-            } else {
-                self.status
-                    .update_model(&model.name, status::ComponentStatus::Initializing);
-                self.load_model(model).await;
-            }
-        }
-
-        // Remove models that are no longer in the app
-        for model in &current_app.models {
-            if !new_app.models.iter().any(|m| m.name == model.name) {
-                self.status
-                    .update_model(&model.name, status::ComponentStatus::Disabled);
-                self.remove_model(model).await;
-            }
-        }
-    }
-
-    /// Update views based on changed between the current and new app.
-    /// This function will update views that have changed, and remove views that are no longer in the app.
-    /// It will also update views that have dependencies that have changed.
-    fn apply_view_diff(&self, current_app: &Arc<App>, new_app: &Arc<App>) {
-        let valid_views = Self::get_valid_views(new_app, LogErrors(true));
-        let existing_views = Self::get_valid_views(current_app, LogErrors(false));
-
-        let views_that_changed = valid_views
-            .iter()
-            .filter_map(|v| {
-                let old_v = existing_views.iter().find(|vv| v.name == vv.name)?;
-                if old_v == v {
-                    None
-                } else {
-                    Some(v.name.clone())
-                }
-            })
-            .collect_vec();
-
-        // Remove views that are no longer in the app
-        for view in &current_app.views {
-            if !new_app.views.iter().any(|v| v.name == view.name) {
-                let view = match View::try_from(view.clone()) {
-                    Ok(v) => v,
-                    Err(e) => {
-                        tracing::error!("Could not remove view {}: {e}", view.name);
-                        continue;
-                    }
-                };
-                self.status
-                    .update_view(&view.name, status::ComponentStatus::Disabled);
-                self.remove_view(&view);
-            }
-        }
-
-        // Get ordering of views to load, including those unchanged but with dependencies that have changed
-        // If we can't determine the order, we'll just load the views in the order they are in the app
-        let affected_views_in_order_of_dependencies = match valid_views
-            .iter()
-            .map(|v| {
-                let Some(statement) =
-                    DFParser::parse_sql_with_dialect(v.sql.as_str(), &PostgreSqlDialect {})
-                        .boxed()?.pop_front() else {
-                            return Err(Box::<dyn std::error::Error + Send + Sync>::from(format!("no statements found in view {}", v.name)));
-                        };
-
-                let deps = view::get_dependent_table_names(&statement);
-                Ok((v.name.clone(), deps))
-            })
-            .collect::<Result<HashMap<TableReference, Vec<TableReference>>, _>>()
-        {
-            Err(e) => {
-                tracing::warn!("Unable to determine order to update views: {e}. Will still attempt to update views.");
-                None
-            }
-            Ok(deps) => construct_effected_in_topological_order(deps,&views_that_changed ),
-        }.unwrap_or(valid_views.iter().map(|v| v.name.clone()).collect());
-
-        for view_name in affected_views_in_order_of_dependencies {
-            if let Some(view) = valid_views.iter().find(|v| v.name == view_name) {
-                if existing_views.iter().any(|v| v.name == view.name) {
-                    // Update view even if unchanged, as it may have dependencies that have changed
-                    self.update_view(view);
-                } else {
-                    self.status
-                        .update_view(&view.name, status::ComponentStatus::Initializing);
-                    let _ = self.load_view(view);
-                }
-            }
-        }
-    }
-
-    async fn apply_dataset_diff(&self, current_app: &Arc<App>, new_app: &Arc<App>) {
-        let valid_datasets = Self::get_valid_datasets(new_app, LogErrors(true));
-        let existing_datasets = Self::get_valid_datasets(current_app, LogErrors(false));
-
-        for ds in valid_datasets {
-            if let Some(current_ds) = existing_datasets.iter().find(|d| d.name == ds.name) {
-                if ds != *current_ds {
-                    self.update_dataset(ds).await;
-                }
-            } else {
-                self.status
-                    .update_dataset(&ds.name, status::ComponentStatus::Initializing);
-                self.load_dataset(ds).await;
-            }
-        }
-
-        // Remove datasets that are no longer in the app
-        for ds in &current_app.datasets {
-            if !new_app.datasets.iter().any(|d| d.name == ds.name) {
-                let ds = match Dataset::try_from(ds.clone()) {
-                    Ok(ds) => ds,
-                    Err(e) => {
-                        tracing::error!("Could not remove dataset {}: {e}", ds.name);
-                        continue;
-                    }
-                };
-                self.status
-                    .update_dataset(&ds.name, status::ComponentStatus::Disabled);
-                self.remove_dataset(&ds).await;
-            }
-        }
-    }
-
-    async fn apply_catalog_diff(&self, current_app: &Arc<App>, new_app: &Arc<App>) {
-        let valid_catalogs = Self::get_valid_catalogs(new_app, LogErrors(true));
-        let existing_catalogs = Self::get_valid_catalogs(current_app, LogErrors(false));
-
-        for catalog in &valid_catalogs {
-            if let Some(current_catalog) = existing_catalogs.iter().find(|c| c.name == catalog.name)
-            {
-                if catalog != current_catalog {
-                    // It isn't currently possible to remove catalogs once they have been loaded in DataFusion. `load_catalog` will overwrite the existing catalog.
-                    self.load_catalog(catalog).await;
-                }
-            } else {
-                self.status
-                    .update_catalog(&catalog.name, status::ComponentStatus::Initializing);
-                self.load_catalog(catalog).await;
-            }
-        }
-    }
-
-    async fn start_pods_watcher(&self) -> notify::Result<()> {
-        let mut pods_watcher = self.pods_watcher.write().await;
-        let Some(mut pods_watcher) = pods_watcher.take() else {
-            return Ok(());
-        };
-        let mut rx = pods_watcher.watch()?;
-
-        while let Some(new_app) = rx.recv().await {
-            let mut app_lock = self.app.write().await;
-            if let Some(current_app) = app_lock.as_mut() {
-                let new_app = Arc::new(new_app);
-                if *current_app == new_app {
-                    continue;
-                }
-
-                tracing::debug!("Updated pods information: {:?}", new_app);
-                tracing::debug!("Previous pods information: {:?}", current_app);
-
-                self.apply_catalog_diff(current_app, &new_app).await;
-                self.apply_dataset_diff(current_app, &new_app).await;
-                self.apply_view_diff(current_app, &new_app);
-                self.apply_model_diff(current_app, &new_app).await;
-
-                *current_app = new_app;
-            } else {
-                *app_lock = Some(Arc::new(new_app));
-            }
-        }
-
-        Ok(())
-    }
-
-    pub async fn init_results_cache(&self) {
-        let app = self.app.read().await;
-        let Some(app) = app.as_ref() else { return };
-
-        let cache_config = &app.runtime.results_cache;
-
-        if !cache_config.enabled {
-            return;
-        }
-
-        match QueryResultsCacheProvider::try_new(
-            cache_config,
-            Box::new([SPICE_RUNTIME_SCHEMA.into(), "information_schema".into()]),
-        ) {
-            Ok(cache_provider) => {
-                tracing::info!("Initialized results cache; {cache_provider}");
-                self.datafusion().set_cache_provider(cache_provider);
-            }
-            Err(e) => {
-                tracing::warn!("Failed to initialize results cache: {e}");
-            }
-        };
-    }
-
-    pub async fn init_task_history(&self) -> Result<()> {
-        let app = self.app.read().await;
-
-        if let Some(app) = app.as_ref() {
-            if !app.runtime.task_history.enabled {
-                tracing::debug!("Task history is disabled via configuration.");
-                return Ok(());
-            }
-        }
-
-        let (retention_period_secs, retention_check_interval_secs) = match app.as_ref() {
-            Some(app) => (
-                app.runtime
-                    .task_history
-                    .retention_period_as_secs()
-                    .map_err(|e| Error::UnableToTrackTaskHistory {
-                        source: task_history::Error::InvalidConfiguration { source: e }, // keeping the spicepod detached but still want to return snafu errors
-                    })?,
-                app.runtime
-                    .task_history
-                    .retention_check_interval_as_secs()
-                    .map_err(|e| Error::UnableToTrackTaskHistory {
-                        source: task_history::Error::InvalidConfiguration { source: e },
-                    })?,
-            ),
-            None => (
-                task_history::DEFAULT_TASK_HISTORY_RETENTION_PERIOD_SECS,
-                task_history::DEFAULT_TASK_HISTORY_RETENTION_CHECK_INTERVAL_SECS,
-            ),
-        };
-
-        match task_history::TaskSpan::instantiate_table(
-            self.status(),
-            retention_period_secs,
-            retention_check_interval_secs,
-        )
-        .await
-        {
-            Ok(table) => self
-                .df
-                .register_runtime_table(
-                    TableReference::partial(
-                        SPICE_RUNTIME_SCHEMA,
-                        task_history::DEFAULT_TASK_HISTORY_TABLE,
-                    ),
-                    table,
-                )
-                .context(UnableToCreateBackendSnafu),
-            Err(source) => Err(Error::UnableToTrackTaskHistory { source }),
-        }
-    }
-}
-
-pub struct RegisterDatasetContext {
-    data_connector: Arc<dyn DataConnector>,
-    federated_read_table: FederatedTable,
-    source: String,
-    accelerated_table: Option<AcceleratedTable>,
 }
 
 #[must_use]
